split installer
